kalman_tools/tools/arch/kalman_photo_app/requirements.txt
Pillow
PyYAML
numpy
scipy